Introduce type aliasing
Provides a simple way to allow type aliases for serialized types. Instead of
relying on the AssemblyQualifiedName, an arbitrary user-defined string can
be registered as a substitute.
This allows interfacing with software not under control of the same developer
without having to mimick the namespace structure on the far end of the
protocol. Also, this allows a sane way of serializing types in the presence of
code obfuscation (a necessary evil in the world of introspectable languages),
especially since obfuscated names change from one obfuscator run to the next.